tripleo_nova_compute: add for standalone roles
TBD: group "flat tasks" by virtual steps to match the expected
tripleo stepX filters. So that when there is no step data inputs
(i.e. it is being deployed outside of t-h-t), the tasks file use
a flat execution model. And if it has steps data passed in via role
vars, consequent filtered include_tasks to be picked up based on that.

Co-Authored-By: James Slagle <jslagle@redhat.com>
Change-Id: I69b135fd76883498d01e51851a337970a7987454
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>